Porytiles2 MVP: palette packing, C parser, animations, and error reporting (#188)
## Summary

Major development push toward the Porytiles2 MVP. This PR introduces
several foundational subsystems and significant improvements across the
codebase.

  ### Palette Packing System
- Implemented full palette packing pipeline: `PackableTile`,
`PackedPalette`, `PalettePool`, `PrefilledPalette`, and
  `PalettePacker` domain service
- Added two packing strategies: **BestFusionStrategy** and
**OverloadAndRemoveStrategy**
- Integrated palette packing into `PrimaryTilesetCompiler` with support
for optimize/locked modes
  - Added palette validation and palette hints

  ### C Parser & Improved Error Reporting
- Built a lightweight C parsing module (lexer + parser) for reading
header defines and behavior maps
- Created `FileHighlightPrinter` utility for rich, source-location-aware
error messages with column highlighting
- Integrated across `HeaderDefineProvider`, `HeaderBehaviorMapProvider`,
and `JascPalLoader`

  ### Animation Support
  - Added basic animation domain types and tileset component integration
  - Introduced key frame concept and key frame mangling

  ### Other Notable Changes
- Refactored tileset artifact handlers to use artifact-specific methods
  - Added `AttributesCsvLoader` for metatile attribute CSV parsing
  - Replaced `unsigned int` with `std::size_t` across the codebase
  - Added wildcard JASC palette loading/saving support
  - Added YAML config key verification and various config updates

+++ Porytiles2/tests/unit/domain/models/canonical_pixel_tile_test.cpp
@@ -18,7 +18,7 @@ TEST(CanonicalPixelTileTests, ShouldFindCanonicalRepresentation)
     PixelTile<IndexPixel> tile{};
     for (std::size_t row = 0; row < 8; ++row) {
         for (std::size_t col = 0; col < 8; ++col) {
-            tile.set(row, col, IndexPixel{static_cast<unsigned int>(row + col + 1)});
+            tile.set(row, col, IndexPixel{row + col + 1});
         }
     }
 
@@ -95,7 +95,7 @@ TEST(CanonicalPixelTileTests, ShouldProduceConsistentResults)
     // Create a test tile
     PixelTile<IndexPixel> tile{};
     for (std::size_t i = 0; i < 64; ++i) {
-        tile.set(i, IndexPixel{static_cast<unsigned int>(i % 16)});
+        tile.set(i, IndexPixel{i % 16});
     }
 
     // Create multiple canonical tiles from the same source
